+++ subiquity/ui/views/filesystem/tests/test_guided.py
@@ -3,10 +3,11 @@ from unittest import mock
 
 import urwid
 
+from subiquitycore.testing import view_helpers
+
 from subiquity.controllers.filesystem import FilesystemController
 from subiquity.ui.views.filesystem.guided import GuidedFilesystemView
 
-from subiquity.ui.views.tests import helpers
 
 
 class GuidedFilesystemViewTests(unittest.TestCase):
@@ -17,7 +18,7 @@ class GuidedFilesystemViewTests(unittest.TestCase):
 
     def test_initial_focus(self):
         view = self.make_view()
-        focus_path = helpers.get_focus_path(view)
+        focus_path = view_helpers.get_focus_path(view)
         for w in reversed(focus_path):
             if isinstance(w, urwid.Button) and w.label == "Use An Entire Disk":
                 return
@@ -26,18 +27,18 @@ class GuidedFilesystemViewTests(unittest.TestCase):
 
     def test_click_guided(self):
         view = self.make_view()
-        button = helpers.find_button_matching(view, "^Use An Entire Disk$")
-        helpers.click(button)
+        button = view_helpers.find_button_matching(view, "^Use An Entire Disk$")
+        view_helpers.click(button)
         view.controller.guided.assert_called_once_with()
 
     def test_click_manual(self):
         view = self.make_view()
-        button = helpers.find_button_matching(view, "^Manual$")
-        helpers.click(button)
+        button = view_helpers.find_button_matching(view, "^Manual$")
+        view_helpers.click(button)
         view.controller.manual.assert_called_once_with()
 
     def test_click_back(self):
         view = self.make_view()
-        button = helpers.find_button_matching(view, "^Back$")
-        helpers.click(button)
+        button = view_helpers.find_button_matching(view, "^Back$")
+        view_helpers.click(button)
         view.controller.cancel.assert_called_once_with()
