+++ subiquity/ui/views/tests/test_welcome.py
@@ -3,11 +3,11 @@ from unittest import mock
 
 import urwid
 
+from subiquitycore.testing import view_helpers
+
 from subiquity.controllers.welcome import WelcomeController
 from subiquity.models.locale import LocaleModel
 from subiquity.ui.views.welcome import WelcomeView
-
-from subiquity.ui.views.tests import helpers
 
 
 class WelcomeViewTests(unittest.TestCase):
@@ -22,8 +22,8 @@ class WelcomeViewTests(unittest.TestCase):
         # Clicking the button for a language calls "switch_language"
         # on the model and "done" on the controller.
         view = self.make_view_with_languages([('code', 'lang', 'native')])
-        but = helpers.find_button_matching(view, "^native$")
-        helpers.click(but)
+        but = view_helpers.find_button_matching(view, "^native$")
+        view_helpers.click(but)
         view.model.switch_language.assert_called_once_with("code")
         view.controller.done.assert_called_once_with()
 
@@ -34,7 +34,7 @@ class WelcomeViewTests(unittest.TestCase):
             ('code1', 'lang1', 'native1'),
             ('code2', 'lang2', 'native2'),
             ])
-        for w in reversed(helpers.get_focus_path(view)):
+        for w in reversed(view_helpers.get_focus_path(view)):
             if isinstance(w, urwid.Button):
                 self.assertEqual(w.label, "native1")
                 break
